+++ packages/0x.js/test/utils/constants.ts
@@ -8,4 +8,5 @@ export const constants = {
     KOVAN_RPC_URL: 'https://kovan.infura.io',
     ROPSTEN_RPC_URL: 'https://ropsten.infura.io',
     ZRX_DECIMALS: 18,
+    GAS_ESTIMATE: 500000,
 };