+++ packages/0x.js/test/utils/web3_factory.ts
@@ -7,7 +7,8 @@ import * as Web3 from 'web3';
 import ProviderEngine = require('web3-provider-engine');
 import RpcSubprovider = require('web3-provider-engine/subproviders/rpc');
 
-import {EmptyWalletSubProvider} from '../../src/subproviders/empty_wallet_subprovider';
+import {EmptyWalletSubprovider} from './subproviders/empty_wallet_subprovider';
+import {FakeGasEstimateSubprovider} from './subproviders/fake_gas_estimate_subprovider';
 
 import {constants} from './constants';
 
@@ -22,8 +23,9 @@ export const web3Factory = {
         const provider = new ProviderEngine();
         const rpcUrl = `http://${constants.RPC_HOST}:${constants.RPC_PORT}`;
         if (!hasAddresses) {
-            provider.addProvider(new EmptyWalletSubProvider());
+            provider.addProvider(new EmptyWalletSubprovider());
         }
+        provider.addProvider(new FakeGasEstimateSubprovider(constants.GAS_ESTIMATE));
         provider.addProvider(new RpcSubprovider({
             rpcUrl,
         }));