+++ packages/website/ts/containers/zero_ex_js_documentation.tsx
@@ -116,6 +116,7 @@ const docsInfoConfig: DocsInfoConfig = {
         'MethodOpts',
         'ValidateOrderFillableOpts',
         'OrderTransactionOpts',
+        'TransactionOpts',
         'ContractEventArg',
         'LogEvent',
         'LogEntry',